TS-OVER: separate UserForm with abstract class View
