+++ src/View/UserForm.ts
@@ -1,19 +1,7 @@
-import { User } from '../Models/User';
+import { User, UserProps } from '../Models/User';
+import { View } from './View';
 
-export class UserForm {
-    constructor(
-        public parent: Element,
-        public model: User
-    ) {
-        this.bindModel();
-    }
-
-    bindModel(): void {
-        this.model.on('change', () => {
-            this.render();
-        });
-    }
-
+export class UserForm extends View<User, UserProps> {
     eventsMap(): { [key: string]: () => void } {
         return {
             'click:.set-age': this.onSetAgeClick,
@@ -43,25 +31,5 @@ export class UserForm {
                 <button class="update-name">Update name</button>
             </div>
         `;
-    }
-
-    bindEvents(fragment: DocumentFragment): void {
-        const eventsMap = this.eventsMap();
-        for (const eventKey in eventsMap) {
-            if (eventsMap.hasOwnProperty(eventKey)) {
-                const [eventName, selector] = eventKey.split(':');
-                fragment.querySelectorAll(selector).forEach(element => {
-                    element.addEventListener(eventName, eventsMap[eventKey])
-                })
-            }
-        }
-    }
-
-    render(): void {
-        this.parent.innerHTML = '';
-        const templateElement = document.createElement('template');
-        templateElement.innerHTML = this.template();
-        this.bindEvents(templateElement.content);
-        this.parent.append(templateElement.content);
     }
 }
