+++ src/View/View.ts
@@ -1,0 +1,39 @@
+import { Model } from '../Models/Model';
+
+export abstract class View<T extends Model<K>, K> {
+    constructor(
+        public parent: Element,
+        public model: T
+    ) {
+        this.bindModel();
+    }
+
+    abstract eventsMap(): { [key: string]: () => void };
+    abstract template(): string;
+
+    bindModel(): void {
+        this.model.on('change', () => {
+            this.render();
+        });
+    }
+
+    bindEvents(fragment: DocumentFragment): void {
+        const eventsMap = this.eventsMap();
+        for (const eventKey in eventsMap) {
+            if (eventsMap.hasOwnProperty(eventKey)) {
+                const [eventName, selector] = eventKey.split(':');
+                fragment.querySelectorAll(selector).forEach(element => {
+                    element.addEventListener(eventName, eventsMap[eventKey])
+                })
+            }
+        }
+    }
+
+    render(): void {
+        this.parent.innerHTML = '';
+        const templateElement = document.createElement('template');
+        templateElement.innerHTML = this.template();
+        this.bindEvents(templateElement.content);
+        this.parent.append(templateElement.content);
+    }
+}
